Redesigned scraper
I should have split this up better. But heres exactly what changed:

-Rename
-Now, the crontab can have additional tags added to it that specify time + frequency of indexing & scraping
-Fetch from the most recent unix timestamp folder name in the data/slack dir
-nuke scrape process if it fails in the filesystem for that time
-index from the most recent unix time in file system to present
-cleaned up threading code to better fit the conversations_replies docs
